WIP: Use snowball stem on populate
When populating Places to the db, use snowball stem version of name to figure out
place is already in the db. This leads to 6350 rows where 2625 had no region  which is about same as with previous Levenshtein implementation. Aftwerwards further combining places and adding regions from same stem Places already in db, only 1844 places still missed region. For now this generated duplicates to the db which is not good. Maybe run additional stem process during population?