Kairatools test data setup
This is not perfect. There seems to be a problem where if we try to reset
the connection to the database to different user, tests take a long time to complete.
It's almost like each test waits some timeout before continuing?  For now this problem is
circumvented by only initializing db_connection once with postgre (admin) user and then running all
tests with it. However, it would be better to connect to the database in Flask application with some other
user since in production the application will not be admin user.