Disable triggers during the mass updates
This should speed up the population a bit and prevent generating useless
logs to audit table. That table is meant for tracking manual changes, not mass
updates. However, for now this code won't work since the trigger disabling requires
user (kaira) to have owner privileges to the table of the trigger.