Remove audit table creation migration
Do the unthinkable and change migration history to remove the audit table
setups. This should be fine since at this point of project we can still recreate the
database from scratch even in production and get rid of useless audit stuff.

If we judge later that we need row level auditing of human made changes, we
can introduce the audit tables back.