Change the spouse population mapping
The json format changed a bit after recent Kaira's Spouse related
fix. Also added a small note to readme about the local connections of the
postgres. For development environment it is probably easier to just use
trusted connections rather than deal with pgpass files but it is up
to you how you do it.