Add logic to form families out of linked data
Using the linked data, this logic creates MiKARELIAn families.
Since not every Katihalian has a link in the MiKARELIAn database,
a family may have Katihalians without a link, because the
researchers want to have data about number of sisters, brothers,
birth order, et cetera. Note that this tool can not output a family
that does not have a single member with a link in the MiKARELIAn
data.