Add DataFormatter to massage the data into its final form
This tool massages the data into a format that we can write
to disk and then use for populating the database. The format
is basically just a list of namedtuples, with those namedtuples
containing all the relevant data we are interested in from the
Katiha side of things, as well as their linked MiKARELIAn person's
kaira_id and their family_id.