Add command to run linked data population
Updated the population code in main.py to use argparse instead
of sys.argv and removed some dead code to run population from
testset.json. The code would never execute because it was set
up to execute if file was not specified, but the way argparse
is set up is that it requires file to be specified.