Make man, woman -> primary, spouse in Marriage
There are cases where people's sex can not reliably be
determined but they and their spouse are still populated
to the database. Because they do not have a sex, however,
their marriage does not get a row in the Marriage table.
With this change, we can populate the marriages without
worrying about manId and womanId being correct.